% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSurvey-package.R
\docType{package}
\name{BiocSurvey-package}
\alias{BiocSurvey}
\alias{BiocSurvey-package}
\title{BiocSurvey: Analysis of Community Survey Results}
\description{
Report and raw data of Bioconductor Community Survey conducted in 2019.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sa-lee/BiocSurvey}
  \item Report bugs at \url{https://github.com/sa-lee/BiocSurvey/issues}
}

}
\author{
\strong{Maintainer}: Stuart Lee \email{lee.s@wehi.edu.au}

}
\keyword{internal}
