% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSurvey-package.R
\docType{data}
\name{questions}
\alias{questions}
\title{Questions asked in 2019 Bioc Community Survey}
\format{A tibble with 40 rows and 1 column with details of each question
\describe{
  \item{question}{the survey question details}
}}
\usage{
questions
}
\description{
Questions asked in 2019 Bioc Community Survey
}
\keyword{datasets}
