% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSurvey-package.R
\docType{data}
\name{responses}
\alias{responses}
\title{Responses to 2019 Bioc Community Survey from Google Forms}
\format{A tibble with 117 rows with 41 Variables
\describe{
  \item{Timestamp}{the timestamp of response from},
  \item{...}{the remaining questions, see `questions` for details}
}}
\usage{
responses
}
\description{
Responses to 2019 Bioc Community Survey from Google Forms
}
\keyword{datasets}
