% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSurvey-package.R
\docType{data}
\name{valid_langs}
\alias{valid_langs}
\title{Valid progamming languages}
\format{A character vector of length 494, containing programming
languages used to check the responses to which programming languages
do you use?}
\usage{
valid_langs
}
\description{
Valid progamming languages
}
\keyword{datasets}
